refactor(foundation): extract mockforge-foundation with Error/Result/EncryptionError
Phase 6 / A1: Foundation crate for shared types with zero mockforge-* dependencies.

Prerequisite: unified the duplicate EncryptionError enums in mockforge-core.
Previously src/encryption.rs had a simple 6-variant EncryptionError and
src/encryption/errors.rs had a richer 17-variant EncryptionError. Both types
coexisted with no conversion between them, making any Error extraction blocked
on the #[from] impl ambiguity.

EncryptionError variant mapping (51 call sites migrated in src/encryption.rs):
- Encryption(s) → CipherOperationFailed { message: s }
- Decryption(s) → CipherOperationFailed { message: s }
- InvalidKey(s) → InvalidKey { message: s }
- InvalidCiphertext(s) → InvalidCiphertext { message: s }
- KeyDerivation(s) → KeyDerivationFailed { message: s }
- Generic unchanged

After unification:
- crates/mockforge-foundation/ created with Error/Result + EncryptionError/EncryptionResult
- mockforge-foundation has zero mockforge-* dependencies (sits at bottom of graph)
- mockforge-core depends on foundation and re-exports via src/error.rs and
  src/encryption/errors.rs (thin shims for backwards compat)
- scripting feature in core forwards to foundation (rquickjs::Error #[from])

Verification:
- 18 foundation tests passing
- 107 core encryption tests passing
- cargo check --workspace clean
