+++ crates/mockforge-core/Cargo.toml
@@ -19,6 +19,7 @@ rustdoc-args = ["--cfg", "docsrs"]
 missing_docs = "deny"
 
 [dependencies]
+mockforge-foundation = { path = "../mockforge-foundation", default-features = false }
 serde = { workspace = true }
 serde_json = { workspace = true }
 serde_yaml = { workspace = true }
@@ -103,7 +104,7 @@ windows = { version = "0.62", features = ["Win32_Security_Credentials"] }
 
 [features]
 default = ["data", "ai", "contracts", "workspace-mgmt", "voice", "advanced", "scripting"]
-scripting = ["rquickjs"]
+scripting = ["rquickjs", "mockforge-foundation/scripting"]
 parallel-routes = ["rayon"]
 schema = ["schemars"]
 database = ["sqlx"]
